新需求前提交git

+++ src/main/java/xyz/soulspace/cinemaline/dto/FilmSimpleDTO.java
@@ -2,11 +2,13 @@ package xyz.soulspace.cinemaline.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class FilmSimpleDTO {
     private int id;
     private String title;
-    private double score;
+    private BigDecimal score;
     private String filmImg;
 }
